requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-moment
flask-mail
pyjwt
numpy
stripe
redis
WTForms
SQLAlchemy
email_validator
psycopg2
gunicorn